Facilitate progress updates
This has been modified to facilitate NSProgress updates. Key to this
was exposing `NSURLSessionTask`, so that we could get the `NSProgress`
from the manager for the task associated with this request. In the
process of exposing the task, the `AFHTTPSessionOperation` was
simplified.

Also added some comments to the headers.

Also made `AFHTTPSessionOperation` initializer non-optional (I’m not
sure why it ever was, as it always returned a value)